requirements.txt
adamp==0.3.0
asttokens
attrs
blessed==1.20.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.4
comm
contourpy==1.0.7
cycler==0.11.0
debugpy
decorator
defusedxml
easing-functions==1.0.4
entrypoints
executing
fastjsonschema
filelock==3.12.1
fonttools==4.40.0
fsspec==2023.6.0
gpustat==1.1
huggingface-hub==0.16.4
idna==3.4
imageio==2.31.1
importlib-metadata
importlib-resources==5.12.0
ipdb==0.13.4
ipykernel
ipython
jedi==0.18.2
Jinja2
joblib==1.2.0
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab-pygments
kiwisolver==1.4.4
lit==16.0.5.post0
lmdb==1.0.0
MarkupSafe
matplotlib==3.7.1
matplotlib-inline
mistune
mpmath==1.3.0
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.5.1.48
packaging
pandas==1.2.1
pandocfilters
parso
pexpect
pickleshare
Pillow==9.5.0
pkgutil_resolve_name
prompt-toolkit==3.0.38
psutil
ptyprocess
pure-eval
Pygments==2.15.1
pyparsing==3.0.9
python-dateutil
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq
rdkit==2023.3.3
referencing
requests==2.31.0
rpds-py
safetensors==0.3.1
scikit-image==0.18.2
scikit-learn==1.2.2
scipy==1.6.0
six
soupsieve
stack-data
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
timm==0.9.2
tinycss2
torch==2.0.1
torchvision==0.15.2
tornado
tqdm==4.56.0
traitlets
triton==2.0.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth
webencodings
zipp==3.15.0
